refactor: move block merging logic into simulator
- Move full `BlockMerger` responsibility from auctioneer to simulator
- Simplify auctioneer to bid ranking + payload storage only
- Trigger merges internally after validation instead of external RPC
- Eliminate auctioneer → simulator ExecutionPayload roundtrip for merges
- Rename auctioneer `BlockMerger` → `MergedBidCache` (cache-only)
- Subscribe simulator to `TopBidForSim` spine queue for base block tracking
- `ValidatedMergeCandidate` now carries full payload from validation success path

Fixes #438
